review: NewspaperTabHeroCopy renamed to newspaperTabHeroData including case change
