+++ support-frontend/assets/pages/paper-subscription-landing/helpers/newsPaperTabHeroData.tsx
